% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{nameCheck}
\alias{nameCheck}
\title{Name your dataframe properly}
\usage{
nameCheck(dataname, subid, conditions, dates_times, drtIDs, trial,
  mics = NULL, hits, rt)
}
\arguments{
\item{dataname}{The name you would like the combined .csv output data.frame to have.}

\item{subid}{The name or position of the column containing your subject IDs.}

\item{conditions}{The name, names, position, or positions of the column(s) containing your subject IDs.}

\item{dates_times}{The name or position of the column containing your dates or times.}

\item{drtIDs}{The name or position of the column containing your DRT IDs if you have multiple DRTs in one file.}

\item{trial}{The name or position of the column containing your trial count.}

\item{mics}{If you have microphones attached to your DRT, this will read the DRT output}

\item{hits}{The name or position of the column containing your hit count.}

\item{rt}{The name or position of the column containing your Response Times.}
}
\description{
Use this function to name your columns properly based on your experimental design
}
\examples{

#df <- nameCheck(combineCSV(fpath = getwd(), dataname = newdf), "V1",c("V3","V4"), "X1","X3","X4",mics = NULL,"X7","X19")

}
