% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizFunctions.R
\name{vizBySubject}
\alias{vizBySubject}
\title{Vizualize data .csv Files}
\usage{
vizBySubject(dataname, plottype, subid, cond, DV)
}
\arguments{
\item{dataname}{The data.frame object you're going to vizualize}

\item{plottype}{Specify the type of plot you'd like to produce}

\item{subid}{The name of the column for your Subject ID numbers.}

\item{cond}{The name of the column for your condition.}

\item{DV}{Dependent Variable of either HR (Hit Rate) or RT (Reatction Time)}
}
\description{
Use these functions to vizualize Detection Response Task outcome variables according to your subjects, conditions, etc. See Examples for more.
}
\examples{

dat <- combineCSV(fpath = getwd(), dataname = newdf)

}
