% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{combineCSV}
\alias{combineCSV}
\title{Combine .csv Files}
\usage{
combineCSV(fpath, dataname = NULL)
}
\arguments{
\item{fpath}{The direcotry that contains all of the .csv files to be combined.}

\item{dataname}{The name you would like the combined .csv output data.frame to have.}
}
\description{
Use this function to read in a series of .csv files and transform them into one data.frame. Make sure to save the output to a data.frame. See Examples for more.
}
\examples{

dat <- combineCSV(fpath = getwd(), dataname = newdf)

}
